tackle renaming the gem (indentation messed up tracking renames)
