--- lib/sidekiq-transaction-guard.rb
@@ -1,3 +1,0 @@
-# frozen_string_literal: true
-
-require 'sidekiq_transaction_guard'
