--- lib/sidekiq_transaction_guard/version.rb
@@ -1,5 +1,0 @@
-# frozen_string_literal: true
-
-module SidekiqTransactionGuard
-  VERSION = File.read(File.expand_path("../../../VERSION", __FILE__)).chomp.freeze
-end
